Extract Chat-window protocol and host adapter
