+++ tests/frontend/chat_ui_unit.cjs
@@ -4,12 +4,14 @@ const path = require("node:path");
 const vm = require("node:vm");
 const { character, cardFile } = require("./character_card_fixtures.cjs");
 const { FakeLocks } = require("./fake_locks.cjs");
-const { getPackageScripts } = require("./script_order.cjs");
+const { getPackageScripts, getScriptPaths } = require("./script_order.cjs");
 
 const ROOT = path.resolve(__dirname, "..", "..");
 const renderingSource = fs.readFileSync(path.join(ROOT, "ui", "js", "chat-rendering.js"), "utf8");
 const appDataSource = fs.readFileSync(path.join(ROOT, "ui", "js", "app-data.js"), "utf8");
-const chatWindowSource = fs.readFileSync(path.join(ROOT, "ui", "js", "chat-window.js"), "utf8");
+const chatWindowScripts = getScriptPaths()
+    .filter(src => src.startsWith("js/chat-window/") || src === "js/chat-window.js")
+    .map(src => ({ path: src, source: fs.readFileSync(path.join(ROOT, "ui", src), "utf8") }));
 // Ordered from ui/index.html via the shared loader helper (script_order.cjs).
 const chatPackageScripts = getPackageScripts("js/chat");
 const chatPackageFiles = chatPackageScripts.map((script) => script.fileName);
@@ -365,7 +367,9 @@ function makeContext({
     chatPackageSources.forEach((pkgSource, i) => {
         vm.runInContext(pkgSource, context, { filename: `ui/js/chat/${chatPackageFiles[i]}` });
     });
-    if (loadChatWindow) vm.runInContext(chatWindowSource, context, { filename: "ui/js/chat-window.js" });
+    if (loadChatWindow) for (const script of chatWindowScripts) {
+        vm.runInContext(script.source, context, { filename: script.path });
+    }
 
     const api = context.window.LlamaGui.chatUi;
     const mutable = {
