+++ ui/index.html
@@ -1908,6 +1908,8 @@
 <script src="/js/chat/chat-stream.js?v=chat-split-1"></script>
 <script src="/js/chat/chat-history.js?v=chat-split-1"></script>
 <script src="/js/chat/chat-main.js?v=chat-split-1"></script>
+<script src="/js/chat-window/chat-window-protocol.js"></script>
+<script src="/js/chat-window/chat-window-host-adapter.js"></script>
 <script src="/js/chat-window.js?v=chat-window-1"></script>
 <script src="/js/benchmark-ui.js?v=revamp-2"></script>
 <script src="/js/inference-stats.js?v=1"></script>
